refactor(search): complete migration acceptance
